More debug. Cast to int. Try to fix sensors from entering infinite loop.

+++ rccar.py
@@ -63,7 +63,7 @@ class RCCar:
         readings = readings[1:-1]
         readings = readings.split(', ')
 
-        readings = [float(i) for i in readings]
+        readings = [int(i) for i in readings]
 
         # The max value in training is 39, so let's reduce to see
         # what happens.
